yeager_ai/requirements/local.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.15.5
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
dill==0.3.6
dj-stripe==2.7.3
Django==4.1.9
django-allauth==0.54.0
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.1
django-htmx==1.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.12.0
jsonfield==3.1.0
jsonschema==4.17.3
kombu==5.3.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.1
pyOpenSSL==23.2.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
slippers==0.5.0
sqlparse==0.4.4
stripe==4.2.0
tomli==2.0.1
tomlkit==0.11.8
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
unicorn==2.0.1.post1
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
zope.interface==6.0